issue/314 - fix bench parse list
